+++ lib/vector_number/stringifying.rb
@@ -47,7 +47,7 @@ class VectorNumber
     # @return [String]
     # @raise [ArgumentError] if +mult+ is not in {MULT_STRINGS}'s keys
     def value_to_s(unit, coefficient, mult:)
-      if mult.is_a?(Symbol) && !MULT_STRINGS.key?(mult)
+      if !mult.is_a?(String) && !MULT_STRINGS.key?(mult)
         raise ArgumentError, "unknown key :#{mult}", caller
       end