+++ CHANGELOG.md
@@ -6,6 +6,10 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Next]
+**Changed**
+- `VectorNumber.new` now accepts options when initializing from a VectorNumber
+   instead of only copying. Options will be merged.
+
 ## [v0.2.3] — 2024-10-15
 **Fixed**
 - BigDecimal tests are now properly skipped when BigDecimal is not available, instead of always.